+++ changelog.md
@@ -1,6 +1,6 @@
 ## Forestry 2.2.3
 - Add Cathemeral bee activity type, which is a random activity period of 12000 ticks that varies based on position. Check times using /forestry bee cathemeral
-- Fix Tolerant Flyer tooltip
+- Fix Tolerant Flyer tooltip and Cave Tolerant display in analyzer
 
 ## Forestry 2.2.2
 - Fix accidental drop shadow in Tab 4 of Analyzer for researched mutations with a + sign