+++ changelog.md
@@ -3,6 +3,7 @@
 - Fix Tolerant Flyer tooltip and Cave Tolerant display in analyzer
 - Fix bees not working below y=0
 - Fix Bog Earth crash when set to maturity = 3 using a debug stick
+- Add worldgen random to IHive#getPosForHive and IHiveGen#getPosForHive
 
 ## Forestry 2.2.2
 - Fix accidental drop shadow in Tab 4 of Analyzer for researched mutations with a + sign