+++ changelog.md
@@ -4,6 +4,7 @@
 - Fix bees not working below y=0
 - Fix Bog Earth crash when set to maturity = 3 using a debug stick
 - Add worldgen random to IHive#getPosForHive and IHiveGen#getPosForHive
+- Fix rendering glitch with Forestry doors (#158)
 
 ## Forestry 2.2.2
 - Fix accidental drop shadow in Tab 4 of Analyzer for researched mutations with a + sign